부산대 BE_강형원 2단계 - 회원 가입 및 로그인 구현 (#181)
* docs: step2 시작 전 readme 작성

* refactor: entity에 대해 다시 class화 (추후 JPA 적용 고려)

* refactor: getter / setter 정리

* docs: readme 내용 추가

* refactor: AddProductRequestDto Record화

* refactor: ModifyProductRequestDto Record화

* refactor: Exception 계층화 및 Handler 간소화

* docs: 주석 정리

* docs: 안 쓰는 package 정리

* feat: dependency 추가

* feat: Member Entity 추가

* feat: Member 관련 Dto / Controller / Service / Repository Skeleton Code 추가

* feat: MemberService register skeleton 작성 (Custom 예외 포함)

* feat: db 설정

* feat: alreadyRegistered 작성

* feat: registerMember 작성

* feat: login 구현

* refactor: service layer, repository layer 책임 분리

* docs: readme 에 api 관련 내용 기록

* feat: test code 작성

* refactor: controller package 단위 분리

* fix: repository layer not found exception throw로 안전성 추가

* fix: request validation 추가

* docs: readme 추가 작성

* feat: member service 책임 분리를 위한 클래스 생성

* fix: jwt key 숨기기

* feat: jwt provider 작성

* feat: auth service (login) 작성

* refactor: 회원가입 및 로그인 매커니즘 재정비(토큰 생성을 로그인으로 통일)

* refactor: Exception 정비

* feat: Product에 대해 admin 권한 지급

* fix: Product Test code 추가 및 수정, Member Test code 추가

+++ src/main/java/gift/controller/product/ProductController.java
@@ -1,9 +1,10 @@
-package gift.controller;
+package gift.controller.product;
 
-import gift.dto.api.AddProductRequestDto;
-import gift.dto.api.ModifyProductRequestDto;
-import gift.dto.api.ProductResponseDto;
-import gift.service.ProductService;
+import gift.dto.api.product.AddProductRequestDto;
+import gift.dto.api.product.ModifyProductRequestDto;
+import gift.dto.api.product.ProductResponseDto;
+import gift.service.auth.AuthService;
+import gift.service.product.ProductService;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.http.HttpStatus;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,16 +25,20 @@ import org.springframework.web.bind.annotation.RestController;
 public class ProductController {
     
     private final ProductService productService;
+    private final AuthService authService;
     
-    public ProductController(ProductService productService) {
+    public ProductController(ProductService productService, AuthService authService) {
         this.productService = productService;
+        this.authService = authService;
     }
     
     //상품 추가 api
     @PostMapping
     public ResponseEntity<ProductResponseDto> addProduct(
-        @RequestBody @Valid AddProductRequestDto requestDto
+        @RequestBody @Valid AddProductRequestDto requestDto,
+        @RequestHeader("Authorization") String authorizationHeader
     ) {
+        authService.checkPermissonForAdmin(authorizationHeader);
         ProductResponseDto responseDto = productService.addProduct(requestDto);
         return new ResponseEntity<>(responseDto, HttpStatus.CREATED);
     }
@@ -57,8 +63,10 @@ public class ProductController {
     @PutMapping("/{productId}")
     public ResponseEntity<ProductResponseDto> modifyProductWithId(
         @PathVariable(name="productId") Long id,
-        @RequestBody @Valid ModifyProductRequestDto requestDto
+        @RequestBody @Valid ModifyProductRequestDto requestDto,
+        @RequestHeader("Authorization") String authorizationHeader
     ) {
+        authService.checkPermissonForAdmin(authorizationHeader);
         ProductResponseDto responseDto = productService.modifyProductWithId(id, requestDto);
         return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
@@ -67,8 +75,10 @@ public class ProductController {
     @PatchMapping("/{productId}")
     public ResponseEntity<ProductResponseDto> modifyProductInfoWithId(
         @PathVariable(name="productId") Long id,
-        @RequestBody @Valid ModifyProductRequestDto requestDto
+        @RequestBody @Valid ModifyProductRequestDto requestDto,
+        @RequestHeader("Authorization") String authorizationHeader
     ) {
+        authService.checkPermissonForAdmin(authorizationHeader);
         ProductResponseDto responseDto = productService.modifyProductInfoWithId(id,
             requestDto);
         return new ResponseEntity<>(responseDto, HttpStatus.OK);
@@ -77,8 +87,10 @@ public class ProductController {
     //상품 단건 삭제 api
     @DeleteMapping("/{productId}")
     public ResponseEntity<Void> deleteProductWithId(
-        @PathVariable(name="productId") Long id
+        @PathVariable(name="productId") Long id,
+        @RequestHeader("Authorization") String authorizationHeader
     ) {
+        authService.checkPermissonForAdmin(authorizationHeader);
         productService.deleteProductWithId(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
